Factor out implementation header (#2089)
